move pairhash
